requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0

# Machine Learning
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
xgboost>=1.7.0

# Experiment Tracking
mlflow==2.9.2

# Data Visualization
seaborn>=0.12.0
matplotlib>=3.7.0
nbformat>=5.0.0

# Testing
pytest==7.4.3
pytest-cov>=4.0.0
pytest-html>=4.0.0
requests>=2.31.0

# Code Quality
flake8==6.1.0
black>=23.0.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0

# Monitoring
prometheus-client>=0.17.0

# Workflow Orchestration (optional)
# prefect>=2.14.0

# Data Loading
# ucimlrepo>=0.0.3

# Documentation
# pdoc3>=0.10.0

# Additional utilities
pyyaml>=6.0
python-json-logger>=2.0.0
